feat: fix path and move plugins component
