+++ client/src/app/plugins/plugins.component.html
