+++ client/src/app/plugins/plugins.component.scss
